refactor(supabase): move server client into supabase.server
Keep the browser module free of Node fs/env.server so client bundles stay safe.
