+++ app/lib/supabase.admin.server.ts
@@ -1,7 +1,9 @@
 import { createClient } from "@supabase/supabase-js";
+import { getEnv } from "./env.server";
 
-const supabaseUrl = process.env.SUPABASE_URL || process.env.NEXT_PUBLIC_SUPABASE_URL!;
-const supabaseServiceRoleKey = process.env.SUPABASE_SERVICE_ROLE_KEY!;
+const supabaseUrl =
+  getEnv("SUPABASE_URL") || getEnv("NEXT_PUBLIC_SUPABASE_URL")!;
+const supabaseServiceRoleKey = getEnv("SUPABASE_SERVICE_ROLE_KEY")!;
 
 export function createAdminClient() {
   if (!supabaseServiceRoleKey) {
